+++ pkg/kubernetes/app.css
@@ -65,7 +65,3 @@ button.fa {
 .container-fluid.alert-dismissible > button > span.pficon {
     vertical-align: middle;
 }
-
-.container-fluid.alert {
-    /*margin: 40px;*/
-}
